requirements.txt
anyio==3.7.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
Flask==2.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
priority==2.0.0
python-dotenv==1.0.0
sniffio==1.3.0
tomli==2.0.1
urllib3==2.0.4
Werkzeug==2.3.7
wsproto==1.2.0
